LayoutTests/media/media-source/media-source-airplay.html

<!DOCTYPE html>
<html>
<head>
    <title>Media engine change after selecting AirPlay targe</title>
    <script src="mock-media-source.js"></script>
    <script src="../media-file.js"></script>
    <script src="../video-test.js"></script>
    <script>

    var sourceBuffer;
    var requestLength = 200000;
    var nextRequest = 0;
    var totalLength = 100;
    var wirelessPlaybackTargetAvailability = '';

    if (window.internals) {
        internals.initializeMockMediaSource();
        internals.setMockMediaPlaybackTargetPickerEnabled(true);
        internals.setMockMediaPlaybackTargetPickerState('', 'Unknown');
    }

    function start() 
    {
        findMediaElement();

        consoleWrite('** create a MediaSource object, set as video.src');
        run(`source = new MediaSource()`);
        run(`source.addEventListener('sourceopen', startLoad)`);
        run(`video.src = URL.createObjectURL(source)`);

        video.addEventListener('webkitcurrentplaybacktargetiswirelesschanged', currentTargetChanged, true);
        video.addEventListener('webkitplaybacktargetavailabilitychanged', targetAvailabilityChanged, true);
        waitForEventOnce('canplaythrough', canplaythrough);
    }

    function startLoad()
    {
        consoleWrite('<br>** create a source buffer, add an init segment with 1 video track');
        run(`sourceBuffer = source.addSourceBuffer('video/mock; codecs="mock"')`);
        sourceBuffer.addEventListener('update', sourceUpdated);
        nextRequest = 0;

        run(`init = makeAInit(100, [makeATrack(1, 'mock', TRACK_KIND.VIDEO)])`);
        run(`sourceBuffer.appendBuffer(init)`);
        consoleWrite('');
    }

    function sourceUpdated()
    {
        if (nextRequest >= totalLength)
            return;

        sourceBuffer.appendBuffer(makeASample(nextRequest, nextRequest, 1, 1, SAMPLE_FLAG.SYNC));
        ++nextRequest;
    }

    function canplaythrough()
    {
        consoleWrite('<br>** simulate choosing a device from the menu');
        runWithKeyDown(function() {
            if (window.internals) {
                internals.setMockMediaPlaybackTargetPickerEnabled(true);
                internals.setMockMediaPlaybackTargetPickerState('Sleepy TV', 'DeviceAvailable');
                video.webkitShowPlaybackTargetPicker();
            }
        });
    }

    function currentTargetChanged()
    {
        if (video.src.indexOf("/test.") > 0) {
            testExpected('video.webkitCurrentPlaybackTargetIsWireless', true);
            testExpected('wirelessPlaybackTargetAvailability', 'available');
            consoleWrite('');
            endTest();
            return;
        }
            
        consoleWrite('<br>** change video.src to one that supports AirPlay');
        video.src = findMediaFile('video', '../content/test')
    }

    function targetAvailabilityChanged(event)
    {
        wirelessPlaybackTargetAvailability = event.availability;
    }
    
    </script>
</head>
<body onload="start()">
    <video controls></video>
</body>
</html>
